PLG-108: Code refactoring
